Add back ratelimit and cpus overwrite in .data section map
seems using unsigned int type in bpf code resolved an
issue that when cpus set to 10, the overwrite failed
for no error and cpus kept the default value.

+++ examples/c/xdppacket.bpf.c
@@ -55,7 +55,7 @@ struct meta_data {
 };
 
 static volatile unsigned int ratelimit = 10;
-static volatile __u8 numcpus = 2;
+static volatile unsigned int numcpus = 2;
 
 struct ipv4_key {
 	struct   bpf_lpm_trie_key lpm_key;
